requirements.txt
pytesseract
pillow
androidviewclient